Simplify cache storage and Next CLI builds
