+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - Preserve build-time Sharp optimization, responsive candidates, WebP fallbacks, art direction, and static export without a CDN.
 - Add Next.js 16.3 integration fixtures for Turbopack and Webpack builds.
 - Make persistent-cache cleanup work when loaders execute in child processes.
+- Isolate asset storage and garbage collection from image transformation.
+- Run maintenance builds through the target application's public Next.js CLI.
 - Use Turbopack for cleanup builds by default and add `next-img --webpack` for backwards compatibility.
 
 ## 0.6.0
