+++ lib/cli.js
@@ -1,70 +1,45 @@
 const path = require('path')
 const crypto = require('crypto')
 const { existsSync } = require('fs')
-const loader = require('./loader')
-const createNextBuild = require('./next-build')
-
-const base = require.resolve('next', { paths: ['.'] }).replace('/dist/server/next.js', '')
-const { Command } = require(`${base}/dist/compiled/commander`)
+const { Command } = require('commander')
+const assetStore = require('./asset-store')
+const runNextBuild = require('./next-cli')
 
 const program = new Command()
 
 program
+  .name('next-img')
   .argument('[dir]', 'directory of the Next.js application (defaults to current directory)')
-  .option('-h, --help', 'display help for command')
   .option('--webpack', 'rebuild using webpack instead of Turbopack')
   .description('Rebuilds all of the images in the persistent cache')
-  .action((dir, options) => {
-    if (options.help) {
-      console.log(`
-      Description
-        Rebuilds all of the images in the persistent cache
-
-      Usage
-        $ next-img <dir>
-
-        <dir> represents the directory of the Next.js application.
-        If no directory is provided, the current directory is used.
-
-      Options
-        --webpack  Rebuild using webpack instead of Turbopack.
-    `)
-      process.exit(0)
-    }
-
+  .action(async (dir, options) => {
     const resolvedDir = path.resolve(dir || '.')
 
-    // Check if the provided directory exists
     if (!existsSync(resolvedDir)) {
       console.error(`> Directory not found: ${resolvedDir}`)
-      process.exit(1)
+      process.exitCode = 1
+      return
     }
 
-    // Pass some info to the next-img plugin via an ENV variable
-    // to tell it to clear and then rebuild the persistent cache
-    process.env.NEXT_IMG_REBUILD = crypto.randomUUID()
-    process.env.NEXT_IMG_PROJECT_DIR = resolvedDir
-
-    const projectBase = require.resolve('next', { paths: [resolvedDir] }).replace('/dist/server/next.js', '')
-    const nextBuild = createNextBuild(projectBase)
+    const session = crypto.randomUUID()
     const bundler = options.webpack ? 'webpack' : 'turbopack'
-    if (options.webpack) {
-      delete process.env.TURBOPACK
-    } else {
-      process.env.TURBOPACK = '1'
+    const env = {
+      ...process.env,
+      NEXT_IMG_REBUILD: session,
+      NEXT_IMG_PROJECT_DIR: resolvedDir,
     }
-    console.log(`> Rebuilding next-img cache with ${bundler}`)
-    nextBuild(resolvedDir, bundler)
-      .then(async () => {
-        await loader.gc()
-        process.exit(0)
-      })
-      .catch(err => {
-        console.error('')
-        console.error('> Build error occurred')
-        console.error(err)
-        process.exit(1)
-      })
+    delete env.TURBOPACK
+
+    try {
+      console.log(`> Rebuilding next-img cache with ${bundler}`)
+      await runNextBuild({ dir: resolvedDir, bundler, env })
+      await assetStore.gc(session)
+    } catch (error) {
+      console.error('')
+      console.error('> Build error occurred')
+      console.error(error)
+      process.exitCode = 1
+    }
   })
 
-program.parse(process.argv, { from: 'node' })
+program.parseAsync(process.argv)
