+++ lib/loader.js
@@ -1,16 +1,11 @@
 const qs = require('qs')
 const path = require('path')
-const os = require('os')
-const crypto = require('crypto')
-const fs = require('fs/promises')
 const sharp = require('sharp')
 const { green } = require('kleur')
-const { rimraf } = require('rimraf')
-const { mkdirp } = require('mkdirp')
 const deepmerge = require('deepmerge')
 const loaderUtils = require('loader-utils')
-const debug = require('debug')('next-img')
 const queue = require('./queue')
+const assetStore = require('./asset-store')
 
 const has = (obj, prop) => Object.prototype.hasOwnProperty.call(obj, prop)
 const overwriteMerge = (destinationArray, sourceArray, _options) => sourceArray
@@ -37,10 +32,6 @@ module.exports = function loader(buffer) {
 }
 
 module.exports.raw = true // get buffer stream instead of utf8 string
-
-// when using the next-img cli, after the build is complete
-// we call gc to keep only the files that were used by the build
-module.exports.gc = gc
 
 async function load(filePath, buffer) {
   // parse the configuration and combine the loader query config with that passed in via webpack config
@@ -93,13 +84,18 @@ async function load(filePath, buffer) {
         const cacheContent = Buffer.concat([buffer, Buffer.from(JSON.stringify(outputOptions), 'utf8')])
         const cacheKey = createFileName({ template: cacheKeyTpl, format: outputFormat, content: cacheContent, width })
         const res = () => toFormat(img.clone().resize(width, null), outputFormat, outputOptions)
-        const { data, format, ...dimensions } = await cached(res, cacheKey, config, progress)
+        const { data, format, ...dimensions } = await assetStore.cached(
+          () => processImage(res),
+          cacheKey,
+          config,
+          progress,
+        )
 
         // Materialize an importable asset. The generated module imports this
         // file, allowing either webpack or Turbopack to emit it normally.
         const attr = { size, density, width: dimensions.width, height: dimensions.height, format }
         const fileName = createFileName({ template: config.imagesName, content: data, ...attr })
-        const assetPath = await materializeAsset(fileName, data, config)
+        const assetPath = await assetStore.stage(fileName, data, config)
         const asset = addAssetImport(imports, filePath, assetPath)
 
         return { asset, ...attr }
@@ -157,14 +153,6 @@ function addAssetImport(imports, filePath, assetPath) {
   return name
 }
 
-async function materializeAsset(fileName, data, config) {
-  const stageDir = path.resolve(config.assetStageDir)
-  const target = resolveWithin(stageDir, fileName)
-  await mkdirp(path.dirname(target))
-  await writeFileAtomic(target, data)
-  return target
-}
-
 function parseResourceQuery(q, allowedKeys) {
   if (!q) return {}
 
@@ -216,136 +204,6 @@ function coerce(obj) {
       obj[k] = asNumber
       continue
     }
-  }
-}
-
-async function cached(fn, cacheKey, config, progressMessage) {
-  const cacheDir = config.persistentCache
-    ? path.join(config.dir, config.persistentCacheDir)
-    : path.join(config.dir, config.distDir, config.cacheDir)
-  const target = path.join(cacheDir, cacheKey)
-
-  let cachedImage
-  try {
-    const data = await fs.readFile(target)
-    const info = await sharp(data).metadata()
-    cachedImage = { data, width: info.width, height: info.height, format: info.format }
-  } catch {}
-
-  if (cachedImage) {
-    debug(`Cache hit ${cacheKey}`)
-    await recordProcessed(cacheDir, cacheKey, config.rebuildSession)
-    return cachedImage
-  }
-
-  if (config.failOnCacheMiss && !config.rebuildSession) {
-    throw new Error(`Missing an optimised image ${cacheKey}. Make sure to rerun next-img.`)
-  }
-
-  console.log(progressMessage)
-  debug(`Cache miss ${cacheKey}`)
-  const ts = Date.now()
-  const processed = await processImage(fn)
-  await mkdirp(cacheDir)
-  await writeFileAtomic(target, processed.data)
-  await recordProcessed(cacheDir, cacheKey, config.rebuildSession)
-  debug(`Processed ${cacheKey} in ${Date.now() - ts}ms`)
-  return processed
-}
-
-async function gc(session = process.env.NEXT_IMG_REBUILD) {
-  if (!session) {
-    debug('No images found')
-    return
-  }
-
-  const root = getGcRoot(session)
-  let configs
-  try {
-    configs = await fs.readdir(root, { withFileTypes: true })
-  } catch {
-    debug('No images found')
-    return
-  }
-
-  try {
-    for (const entry of configs.filter(entry => entry.isDirectory())) {
-      const configRoot = path.join(root, entry.name)
-      const { cacheDir } = JSON.parse(await fs.readFile(path.join(configRoot, 'config.json'), 'utf8'))
-      const markers = await fs.readdir(path.join(configRoot, 'files'))
-      const processed = new Set(
-        await Promise.all(markers.map(marker => fs.readFile(path.join(configRoot, 'files', marker), 'utf8'))),
-      )
-      await pruneCache(cacheDir, processed)
-    }
-  } finally {
-    await rimraf(root)
-  }
-}
-
-async function recordProcessed(cacheDir, cacheKey, session) {
-  if (!session) return
-
-  const configRoot = path.join(getGcRoot(session), hash(cacheDir))
-  const filesRoot = path.join(configRoot, 'files')
-  await mkdirp(filesRoot)
-  await writeFileAtomic(path.join(configRoot, 'config.json'), JSON.stringify({ cacheDir }))
-  await writeFileAtomic(path.join(filesRoot, hash(cacheKey)), cacheKey)
-}
-
-async function pruneCache(cacheDir, processed) {
-  const parent = path.dirname(cacheDir)
-  await mkdirp(parent)
-  const temp = await fs.mkdtemp(path.join(parent, '.next-img-gc-'))
-
-  try {
-    try {
-      await fs.rename(cacheDir, path.join(temp, 'cache'))
-    } catch (error) {
-      if (error.code !== 'ENOENT') throw error
-      await mkdirp(path.join(temp, 'cache'))
-    }
-
-    await mkdirp(cacheDir)
-    for (const cacheKey of processed) {
-      const source = resolveWithin(path.join(temp, 'cache'), cacheKey)
-      const target = resolveWithin(cacheDir, cacheKey)
-      await mkdirp(path.dirname(target))
-      await fs.rename(source, target)
-    }
-  } finally {
-    await rimraf(temp)
-  }
-}
-
-function getGcRoot(session) {
-  return path.join(os.tmpdir(), 'next-img-gc', hash(session))
-}
-
-function hash(value) {
-  return crypto.createHash('sha256').update(value).digest('hex')
-}
-
-function resolveWithin(root, fileName) {
-  const target = path.resolve(root, fileName)
-  const relative = path.relative(root, target)
-  if (relative.startsWith('..') || path.isAbsolute(relative)) {
-    throw new Error(`next-img generated an unsafe file path: ${fileName}`)
-  }
-  return target
-}
-
-async function writeFileAtomic(target, data) {
-  const temp = `${target}.${process.pid}.${crypto.randomBytes(6).toString('hex')}.tmp`
-  await fs.writeFile(temp, data)
-  try {
-    await fs.rename(temp, target)
-  } catch (error) {
-    if (!['EEXIST', 'EPERM'].includes(error.code)) throw error
-    await fs.rm(target, { force: true })
-    await fs.rename(temp, target)
-  } finally {
-    await fs.rm(temp, { force: true })
   }
 }
