+++ test/asset-store.test.js
@@ -1,0 +1,42 @@
+const fs = require('fs')
+const os = require('os')
+const path = require('path')
+const { default: test } = require('ava')
+const sharp = require('sharp')
+const assetStore = require('../lib/asset-store')
+
+test('garbage collection uses filesystem markers from build workers', async t => {
+  const dir = fs.mkdtempSync(path.join(os.tmpdir(), 'next-img-store-'))
+  const session = `test-${process.pid}-${Date.now()}`
+  const cacheDir = path.join(dir, 'resources')
+  const data = await sharp({
+    create: {
+      width: 8,
+      height: 8,
+      channels: 3,
+      background: { r: 20, g: 40, b: 60 },
+    },
+  })
+    .jpeg()
+    .toBuffer()
+
+  t.teardown(() => fs.rmSync(dir, { recursive: true, force: true }))
+
+  await assetStore.cached(
+    async () => ({ data, width: 8, height: 8, format: 'jpeg' }),
+    'used.jpg',
+    {
+      dir,
+      persistentCache: true,
+      persistentCacheDir: 'resources',
+      rebuildSession: session,
+      failOnCacheMiss: false,
+    },
+    'processing test image',
+  )
+  fs.writeFileSync(path.join(cacheDir, 'unused.jpg'), 'unused')
+
+  await assetStore.gc(session)
+
+  t.deepEqual(fs.readdirSync(cacheDir), ['used.jpg'])
+})
